keep ogg timing when the first audio page is lost, and size buffers safely
- ogg: a lost or damaged first audio page used to pull the whole song
  earlier by the missing audio; the first granule position now says how
  much came before, and silence takes its place.
- ogg: the output buffer was sized from the last granule position alone, so
  a stream whose granules don't start at zero (a cut radio recording) asked
  for up to 64 frames per file byte (326 mb for a 1.2 mb song). it is now
  bounded by the frames the packets actually make.
- wav: a data chunk claiming nearly 2 gb (streamed output) overflowed the
  chunk walk and failed; the size is compared in long arithmetic.
- the frame cap stays within the largest array .net allows, and the decode
  log names missing pages and multichannel wav files accurately.

+++ source/Audio/WavFile.cs
@@ -21,7 +21,7 @@ internal static class WavFile
         {
             int size = BitConverter.ToInt32(b, pos + 4);
             int body = pos + 8;
-            if (size < 0 || body + size > b.Length) size = b.Length - body;
+            if (size < 0 || (long)body + size > b.Length) size = b.Length - body;
             if (b[pos] == id[0] && b[pos + 1] == id[1] && b[pos + 2] == id[2] && b[pos + 3] == id[3]) return (body, size);
             pos = body + size + (size & 1);
         }
@@ -38,7 +38,9 @@ internal static class WavFile
             string id = System.Text.Encoding.ASCII.GetString(b, pos, 4);
             int size = BitConverter.ToInt32(b, pos + 4);
             int body = pos + 8;
-            if (size < 0 || body + size > b.Length) size = b.Length - body;
+            // In long arithmetic: a streamed file can claim a length near 2 GB, which must be
+            // cut to the file rather than wrap around.
+            if (size < 0 || (long)body + size > b.Length) size = b.Length - body;
             if (id == "fmt ")
             {
                 format = BitConverter.ToUInt16(b, body);
